+++ frontend/src/app/features/test/name-input/name-input.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { FormsModule } from '@angular/forms';
+import { ProfileService } from '../../../services';
 
-import { ProfileService } from '../../services/profile.service';
 
 @Component({
   selector: 'app-name-input',
