+++ frontend/src/app/services/index.ts
@@ -1,2 +1,4 @@
 export * from './auth.service';
 export * from './hello.service';
+export * from './profile.service';
+export * from './localstorage.service'
